Update syntax for latest SDL3 version

+++ src/units.h
@@ -2,10 +2,9 @@
 
 namespace Units {
 	/// @brief Screen pixel units, will vary from Game units depending on output window resolution
-	typedef int Pixels;
+	typedef float Pixels;
 	/// @brief Game units (one NES pixel is one Game unit)
-	typedef double Game;
+	typedef float Game;
 	/// @brief Tile units (16 Game units are one Tile unit)
 	typedef int Tile;
-
 };
